feat: move flatten type to shared folder
