+++ packages/ui/react/src/colors/index.ts
@@ -1,2 +1,1 @@
 export * from './getColor'
-export * from './types'
